Leap leap
Check year is a leap year.

Check for the biggest contraints divisible for 400 1st
Check for the second biggest contraints divisible for 100 2nd
Check for the smallest contraints divisible for 4

Notice: follow the info of the problem is truthful; we cannot counter
for the opposite info.

As a result:
Leap year is returned.